+++ src/Symfony/Component/Process/Tests/ProcessBuilderTest.php
@@ -115,4 +115,16 @@ class ProcessBuilderTest extends \PHPUnit_Framework_TestCase
 
         $this->assertContains("second", $proc->getCommandLine());
     }
+
+    public function testShouldEscapeArguments()
+    {
+        $pb = new ProcessBuilder(array('%path%', 'foo " bar'));
+        $proc = $pb->getProcess();
+
+        if (defined('PHP_WINDOWS_VERSION_BUILD')) {
+            $this->assertSame('^%"path"^% "foo "\\"" bar"', $proc->getCommandLine());
+        } else {
+            $this->assertSame("'%path%' 'foo \" bar'", $proc->getCommandLine());
+        }
+    }
 }